Optimized sm_timing for shorter compile time
